requirements.txt
discord.py
asyncio
logging
aiohttp
async-timeout
youtube-dl
